Move PaidAdsFeaturesSection and PaidAdsSetupSections components.
